move the docs package under a commands/ directory
