+++ commands/docs/docs_test.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"path/filepath"
 
-	. "github.com/opencontrol/compliance-masonry/docs"
+	. "github.com/opencontrol/compliance-masonry/commands/docs"
 	"github.com/opencontrol/compliance-masonry/docx"
 	"github.com/opencontrol/compliance-masonry/gitbook"
 
@@ -27,7 +27,7 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check the template method returns an error message when no template is defined",
 			docx.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures"),
 				TemplatePath:   "",
 				ExportPath:     "",
 			},
@@ -37,7 +37,7 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check the template method returns an error message when no template does not exist",
 			docx.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures"),
 				TemplatePath:   "fake",
 				ExportPath:     "",
 			},
@@ -47,8 +47,8 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check that template is created when inputs are correct",
 			docx.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures"),
-				TemplatePath:   filepath.Join("..", "fixtures", "template_fixtures", "test.docx"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures"),
+				TemplatePath:   filepath.Join("..", "..", "fixtures", "template_fixtures", "test.docx"),
 				ExportPath:     "",
 			},
 			nil,
@@ -66,7 +66,7 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check that the gitbook is correctly exported given the fixtures",
 			gitbook.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures"),
 				Certification:  "LATO",
 				MarkdownPath:   "",
 			},
@@ -88,9 +88,9 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check that gitbook is created with markdowns",
 			gitbook.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown"),
 				Certification:  "LATO",
-				MarkdownPath:   filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns"),
+				MarkdownPath:   filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns"),
 			},
 			"",
 			nil,
@@ -99,20 +99,20 @@ var _ = Describe("Doc Tests", func() {
 		Entry(
 			"Check that thre is an error returned when the certification does not exist",
 			gitbook.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown"),
 				Certification:  "LAT",
-				MarkdownPath:   filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns"),
+				MarkdownPath:   filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns"),
 			},
 			"",
-			[]error{fmt.Errorf("Error: `%s` does not exist\nUse one of the following:\nLATO", filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown", "certifications", "LAT.yaml"))},
+			[]error{fmt.Errorf("Error: `%s` does not exist\nUse one of the following:\nLATO", filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown", "certifications", "LAT.yaml"))},
 		),
 
 		Entry(
 			"Check that error is returned when certification argument is not present",
 			gitbook.Config{
-				OpencontrolDir: filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown"),
+				OpencontrolDir: filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown"),
 				Certification:  "",
-				MarkdownPath:   filepath.Join("..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns/"),
+				MarkdownPath:   filepath.Join("..", "..", "fixtures", "opencontrol_fixtures_with_markdown", "markdowns/"),
 			},
 			"",
 			[]error{errors.New("Error: Missing Certification Argument")},
